feat(mgb42-complete-profile-user): create complete user profile for editing
